final tweaks and checks
